requirements.txt
langchain
langchain-community
streamlit
python-dotenv
faiss-cpu
sentence-transformers
tf-keras